Quickly added functionality to remove players from rating.
This makes it so that players that won't play anymore don't end up with
an infinite horizontal line in the history graph. But all statistics is
still saved, as well as the rating history up until the removal.
A removed player that starts playing again will be rated as a completely
new player i.e. will have an initial rating period.
This update is code-wise quite obfuscated, and it is not a very nice
solution, but it is the quickest I came up with for now.